change version to int, implement clock bean

+++ src/main/java/com/blog/myblog/domain/AppUser.java
@@ -54,10 +54,10 @@ public class AppUser implements ConcurrencySafeEntity {
 
     @Column(name = "version")
     @Version
-    private Short version;
+    private int version;
 
     @Override
-    public Short version() {
+    public Integer version() {
         return this.version;
     }
