+++ src/main/java/com/blog/myblog/domain/Article.java
@@ -45,7 +45,7 @@ public class Article implements ConcurrencySafeEntity {
 
     @Column(name = "version")
     @Version
-    private Short version;
+    private int version;
 
     @OneToMany (mappedBy = "article")
     private List<Comment> comments;
@@ -57,7 +57,7 @@ public class Article implements ConcurrencySafeEntity {
     private List<Comment> alerts;
 
     @Override
-    public Short version() {
+    public Integer version() {
         return this.version;
     }
